Rename hyperfuel_source module to fuel_hypersync_source
Match the evm_hypersync_source / svm_hypersync_source naming convention.
The napi HyperfuelClient class name is unchanged, so no JS/ReScript
references need updating.
