+++ packages/cli/src/fuel_hypersync_source/query.rs
